Rename gf_mul() to gf_mul_isal()
When ClickHouse links against OpenSSL instead of BoringSSL [0], then
linking fails due to duplicate symbol gf_mul in OpenSSL and in ISA-L:

  ld.lld-17: error: duplicate symbol: gf_mul
  >>> defined at f_impl64.c:25 (./contrib/openssl/crypto/ec/curve448/arch_64/f_impl64.c:25)
  >>>            f_impl64.c.o:(gf_mul) in archive contrib/openssl-cmake/libcrypto.a
  >>> defined at ec_base.c:48 (./contrib/isa-l/erasure_code/ec_base.c:48)
  >>>            ec_base.c.o:(.text+0x5E0) in archive contrib/isa-l-cmake/lib_isal.a
  clang++-17: error: linker command failed with exit code 1 (use -v to see invocation)
  ninja: build stopped: subcommand failed.

The issue was noted upstream [1], and the upstream fix is to rename the
symbol in OpenSSL (which is only internal, in contrast to ISA-L where it
is in the public API). As far as I see, no FIPS-certified OpenSSL
contains the upstream fix yet, so following this route would make
ClickHouse FIPS-noncompliant (as it changes the source Git hash).
Therefore, this fix rather renames the symbol in ISA-L. This is a
temporary measure until a FIPS-certified OpenSSL with the actual fix
exists.

[0] https://github.com/ClickHouse/ClickHouse/pull/56398
[1] https://github.com/openssl/openssl/issues/21302